Move fetch call into App instead of Main
